+++ sandbox/Sandbox.AspNetCore3/Views/Shared/_Layout.cshtml
@@ -1,0 +1,12 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8" />
+    <title>@ViewData["Title"] - Sandbox.AspNetCore3</title>
+</head>
+<body>
+    <main role="main">
+        @RenderBody()
+    </main>
+</body>
+</html>
