+++ sandbox/Sandbox.AspNetCore3/Views/_ViewImports.cshtml
@@ -1,0 +1,2 @@
+﻿@using Sandbox.AspNetCore3
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
