+++ src/MagicOnion.AspNetCore/Glue/MagicOnionGlueServiceMethodProviderBase.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using Grpc.Core;
+
+namespace MagicOnion.AspNetCore
+{
+    internal class MagicOnionGlueServiceMethodProviderBase
+    {
+        protected ServerServiceDefinition ServerServiceDefinition { get; }
+
+        protected MagicOnionGlueServiceMethodProviderBase(ServerServiceDefinition serverServiceDefinition)
+        {
+            ServerServiceDefinition = serverServiceDefinition;
+        }
+    }
+}
