+++ src/MagicOnion/Server/Hubs/Group.ImmutableArray.cs
@@ -504,7 +504,9 @@ namespace MagicOnion.Server.Hubs
                 }
                 catch (Exception ex)
                 {
+#if !GRPC_DOTNET
                     Grpc.Core.GrpcEnvironment.Logger?.Error(ex, "error occured on write to client, but keep to write other clients.");
+#endif
                 }
             }
         }
@@ -520,7 +522,9 @@ namespace MagicOnion.Server.Hubs
                 }
                 catch (Exception ex)
                 {
+#if !GRPC_DOTNET
                     Grpc.Core.GrpcEnvironment.Logger?.Error(ex, "error occured on write to client, but keep to write other clients.");
+#endif
                 }
             }
         }
