+++ src/MagicOnion/Server/IMagicOnionLogger.cs
@@ -71,35 +71,47 @@ namespace MagicOnion.Server
 
         public void BeginBuildServiceDefinition()
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug(nameof(BeginBuildServiceDefinition));
+#endif
         }
 
         public void EndBuildServiceDefinition(double elapsed)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndBuildServiceDefinition)} elapsed:{elapsed}");
+#endif
         }
 
         public void BeginInvokeMethod(ServiceContext context, byte[] request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeMethod)} type:{MethodTypeToString(context.MethodType)} method:{context.CallContext.Method} size:{request.Length}");
+#endif
         }
 
         public void EndInvokeMethod(ServiceContext context, byte[] response, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeMethod)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{response.Length} elapsed:{elapsed} {msg}");
+#endif
         }
 
         public void WriteToStream(ServiceContext context, byte[] writeData, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(WriteToStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{writeData.Length}");
+#endif
         }
 
         public void ReadFromStream(ServiceContext context, byte[] readData, Type type, bool complete)
         {
             if (context.ServiceType == typeof(EmbeddedServices.MagicOnionEmbeddedHeartbeat)) return;
 
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(ReadFromStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{readData.Length} complete:{complete}");
+#endif
         }
 
         // enum.ToString is slow.
@@ -122,19 +134,24 @@ namespace MagicOnion.Server
 
         public void BeginInvokeHubMethod(StreamingHubContext context, ArraySegment<byte> request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeHubMethod)} method:{context.Path} size:{request.Count}");
-
+#endif
         }
 
         public void EndInvokeHubMethod(StreamingHubContext context, int responseSize, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeHubMethod)} method:{context.Path} size:{responseSize} elapsed:{elapsed} {msg}");
+#endif
         }
 
         public void InvokeHubBroadcast(string groupName, int responseSize, int broadcastGroupCount)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(InvokeHubBroadcast)} size:{responseSize} broadcastGroupCount:{broadcastGroupCount}");
+#endif
         }
     }
 
@@ -150,35 +167,47 @@ namespace MagicOnion.Server
 
         public void BeginBuildServiceDefinition()
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug(nameof(BeginBuildServiceDefinition));
+#endif
         }
 
         public void EndBuildServiceDefinition(double elapsed)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndBuildServiceDefinition)} elapsed:{elapsed}");
+#endif
         }
 
         public void BeginInvokeMethod(ServiceContext context, byte[] request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeMethod)} type:{MethodTypeToString(context.MethodType)} method:{context.CallContext.Method} size:{request.Length} {ToJson(request)}");
+#endif
         }
 
         public void EndInvokeMethod(ServiceContext context, byte[] response, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeMethod)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{response.Length} elapsed:{elapsed} {msg} {ToJson(response)}");
+#endif
         }
 
         public void WriteToStream(ServiceContext context, byte[] writeData, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(WriteToStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{writeData.Length} {ToJson(writeData)}");
+#endif
         }
 
         public void ReadFromStream(ServiceContext context, byte[] readData, Type type, bool complete)
         {
             if (context.ServiceType == typeof(EmbeddedServices.MagicOnionEmbeddedHeartbeat)) return;
 
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(ReadFromStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{readData.Length} complete:{complete} {ToJson(readData)}");
+#endif
         }
 
         string ToJson(byte[] bytes)
@@ -217,18 +246,24 @@ namespace MagicOnion.Server
 
         public void BeginInvokeHubMethod(StreamingHubContext context, ArraySegment<byte> request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeHubMethod)} method:{context.Path} size:{request.Count} {ToJson(request)}");
+#endif
         }
 
         public void EndInvokeHubMethod(StreamingHubContext context, int responseSize, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeHubMethod)} method:{context.Path} size:{responseSize} elapsed:{elapsed} {msg}");
+#endif
         }
 
         public void InvokeHubBroadcast(string groupName, int responseSize, int broadcastGroupCount)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(InvokeHubBroadcast)} size:{responseSize} broadcastGroupCount:{broadcastGroupCount}");
+#endif
         }
     }
 
@@ -252,35 +287,47 @@ namespace MagicOnion.Server
 
         public void BeginBuildServiceDefinition()
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug(nameof(BeginBuildServiceDefinition));
+#endif
         }
 
         public void EndBuildServiceDefinition(double elapsed)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndBuildServiceDefinition)} elapsed:{elapsed}");
+#endif
         }
 
         public void BeginInvokeMethod(ServiceContext context, byte[] request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeMethod)} type:{MethodTypeToString(context.MethodType)} method:{context.CallContext.Method} size:{request.Length} {ToJson(request, type, context)}");
+#endif
         }
 
         public void EndInvokeMethod(ServiceContext context, byte[] response, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeMethod)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{response.Length} elapsed:{elapsed} {msg} {ToJson(response, type, context)}");
+#endif
         }
 
         public void WriteToStream(ServiceContext context, byte[] writeData, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(WriteToStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{writeData.Length} {ToJson(writeData, type, context)}");
+#endif
         }
 
         public void ReadFromStream(ServiceContext context, byte[] readData, Type type, bool complete)
         {
             if (context.ServiceType == typeof(EmbeddedServices.MagicOnionEmbeddedHeartbeat)) return;
 
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(ReadFromStream)} type:{MethodTypeToString(context.MethodType)}  method:{context.CallContext.Method} size:{readData.Length} complete:{complete} {ToJson(readData, type, context)}");
+#endif
         }
 
         string ToJson(byte[] bytes, Type type, ServiceContext context)
@@ -324,18 +371,24 @@ namespace MagicOnion.Server
         }
         public void BeginInvokeHubMethod(StreamingHubContext context, ArraySegment<byte> request, Type type)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(BeginInvokeHubMethod)} method:{context.Path} size:{request.Count} {ToJson(request, type, context.FormatterResolver)}");
+#endif
         }
 
         public void EndInvokeHubMethod(StreamingHubContext context, int responseSize, Type type, double elapsed, bool isErrorOrInterrupted)
         {
             var msg = isErrorOrInterrupted ? "error" : "";
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(EndInvokeHubMethod)} method:{context.Path} size:{responseSize} elapsed:{elapsed} {msg}");
+#endif
         }
 
         public void InvokeHubBroadcast(string groupName, int responseSize, int broadcastGroupCount)
         {
+#if !GRPC_DOTNET
             GrpcEnvironment.Logger.Debug($"{nameof(InvokeHubBroadcast)} size:{responseSize} broadcastGroupCount:{broadcastGroupCount}");
+#endif
         }
     }
